car_rental_campaign/requirements.txt
cassandra-driver
python-dotenv
